+++ CHANGELOG.md
@@ -21,6 +21,8 @@ high state of flux, you're at risk of it changing without notice.
     - add missing instances to `getValidation` (@gcanti)
   - `IOEither`
     - relax `Bifunctor2C` to `Bifunctor2` in `getIOValidation` (@gcanti)
+  - `ReaderEither`
+    - relax `Bifunctor3C` to `Bifunctor3` in `getReaderValidation` (@gcanti)
 
 # 2.5.1